Extract packstream functionality
